+++ lib/src/pages/changelog.dart
@@ -15,6 +15,14 @@ class ChangelogPage extends StatelessWidget {
   _showChangelog() {
     final String changelog = """
 
+## Version 1.1.1
+### **Added** 🚀
+- Added Avatars to Tasks, subtasks, comments and Kanban
+- Changed Logo design to comply with Google Play Policies
+
+### **Fixed**  🔧
+- Kanban now refresh elements after pop out from task page.
+
 ## Version 1.1.0
 ### **Added** 🚀
 - Changelog Section in Side Menu showing this Changelog