+++ lib/src/utils/utils.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:khanos/src/preferences/user_preferences.dart';
 import 'package:khanos/src/utils/theme_utils.dart';
+import 'dart:convert';
+import 'package:crypto/crypto.dart' as crypto;
 
 final _prefs = new UserPreferences();
 ScrollController scrollController = new ScrollController();
@@ -36,7 +38,7 @@ void showChangelog(BuildContext context) {
           scrollable: true,
           title: Text('Welcome!'),
           content: Text(
-              'Welcome to version 1.0.9! You can see what\'s new in the Changelog Section.'),
+              'Welcome to version 1.1.1! You can see what\'s new in the Changelog Section.'),
           actions: <Widget>[
             TextButton(
                 child: Text('View Changes'),
@@ -156,4 +158,13 @@ Widget errorPage(Map<String, dynamic> error) {
       ),
     ],
   );
+}
+
+String getAvatarUrl(String userId, String avatarPath, String size) {
+  String endPoint = _prefs.endpoint;
+  String domain = endPoint.replaceAll("jsonrpc.php", "");
+  String hash = crypto.md5.convert(utf8.encode(avatarPath)).toString();
+  String avatarUrl =
+      '$domain?controller=AvatarFileController&action=image&user_id=$userId&hash=$hash&size=$size';
+  return avatarUrl;
 }